Reorganise files; remove unnecessary table drops
