--- api/application/create-canteen.js
@@ -1,6 +1,0 @@
-require('../database/setup');
-const { Canteen } = require('../database/models');
-
-exports.createCanteen = function(canteen) {
-  return Canteen.create(canteen);
-};
