+++ api/infrastructure/models/user.js
@@ -1,6 +1,6 @@
 // For some reason, removing the unused Sequelize from here causes the tests to break
 const { Sequelize, DataTypes } = require('sequelize');
-const { sequelize } = require('./setup');
+const { sequelize } = require('../postgres-database');
 
 exports.User = sequelize.define('User', {
   id: {
@@ -31,28 +31,5 @@ exports.User = sequelize.define('User', {
   canteenId: {
     type: DataTypes.INTEGER,
     allowNull: false,
-  }
-});
-
-exports.Canteen = sequelize.define('Canteen', {
-  id: {
-    type: DataTypes.INTEGER,
-    primaryKey: true,
-    unique: true,
-    autoIncrement: true,
-    allowNull: false
-  },
-  name: {
-    type: DataTypes.STRING,
-    allowNull: false
-  },
-  // TODO: validate department and sector input
-  department: {
-    type: DataTypes.STRING,
-    allowNull: false
-  },
-  sector: {
-    type: DataTypes.STRING,
-    allowNull: false
   }
 });
