+++ api/infrastructure/repositories/canteen.js
@@ -1,0 +1,6 @@
+require('../postgres-database');
+const { Canteen } = require('../models/canteen');
+
+exports.createCanteen = function(canteen) {
+  return Canteen.create(canteen);
+};
